update: mengubah struktur styling component agar lebih mudah digunakan
