+++ src/assets/less/components/custom-components/TableCard.less
@@ -1,0 +1,20 @@
+.custom-component-table-card {
+
+  border-radius: 10px;
+
+  .table-tools-container {
+    margin-bottom: 30px;
+
+    .button-right {
+      display: flex;
+      justify-content: flex-end;
+
+      button {
+        height: 30px;
+        width: 100%;
+        max-width: 130px;
+        font-size: 13px;
+      }
+    }
+  }
+}
