+++ src/assets/less/components/custom-components/TableDisplay.less
@@ -1,21 +1,4 @@
-.wedding-organizer {
-  border-radius: 10px;
-
-  .table-tools-container {
-    margin-bottom: 30px;
-
-    .button-right {
-      display: flex;
-      justify-content: flex-end;
-
-      button {
-        height: 30px;
-        width: 100%;
-        max-width: 130px;
-        font-size: 13px;
-      }
-    }
-  }
+.custom-component-table-card {
 
   table {
