+++ src/assets/less/components/custom-components/index.less
@@ -1,0 +1,4 @@
+@import "./TableCard.less";
+@import "./TableSearch.less";
+@import "./TableDisplay.less";
+@import "./TableTitle.less";
