+++ src/assets/less/components/index.less
@@ -4,3 +4,4 @@
 @import "./feedback/index.less";
 @import "./navigation/index.less";
 @import "./layout/index.less";
+@import "./custom-components/index.less";
