+++ src/assets/less/pages/index.less
@@ -1,3 +1,2 @@
 @import "error/index.less";
 @import "dashboard/index.less";
-@import "master-data/index.less";
