--- src/assets/less/pages/master-data/index.less
@@ -1,1 +1,0 @@
-@import "./wedding-organizer/index.less";
